+++ src/engine/controls/bpmcontrol.cpp
@@ -5,7 +5,7 @@
 #include "control/controllinpotmeter.h"
 
 #include "engine/enginebuffer.h"
-#include "engine/bpmcontrol.h"
+#include "engine/controls/bpmcontrol.h"
 #include "waveform/visualplayposition.h"
 #include "engine/enginechannel.h"
 #include "engine/enginemaster.h"
