+++ src/engine/controls/bpmcontrol.h
@@ -4,7 +4,7 @@
 #include <gtest/gtest_prod.h>
 
 #include "control/controlobject.h"
-#include "engine/enginecontrol.h"
+#include "engine/controls/enginecontrol.h"
 #include "engine/sync/syncable.h"
 #include "util/tapfilter.h"
