+++ src/engine/controls/clockcontrol.h
@@ -2,7 +2,7 @@
 #define CLOCKCONTROL_H
 
 #include "preferences/usersettings.h"
-#include "engine/enginecontrol.h"
+#include "engine/controls/enginecontrol.h"
 
 #include "track/track.h"
 #include "track/beats.h"
