+++ src/engine/controls/loopingcontrol.cpp
@@ -7,9 +7,9 @@
 #include "control/controlobject.h"
 #include "preferences/usersettings.h"
 #include "control/controlpushbutton.h"
-#include "engine/loopingcontrol.h"
-#include "engine/bpmcontrol.h"
-#include "engine/enginecontrol.h"
+#include "engine/controls/loopingcontrol.h"
+#include "engine/controls/bpmcontrol.h"
+#include "engine/controls/enginecontrol.h"
 #include "util/compatibility.h"
 #include "util/math.h"
 #include "util/sample.h"
