+++ src/engine/controls/loopingcontrol.h
@@ -9,7 +9,7 @@
 #include <QStack>
 
 #include "preferences/usersettings.h"
-#include "engine/enginecontrol.h"
+#include "engine/controls/enginecontrol.h"
 #include "track/track.h"
 #include "track/beats.h"
 #include "control/controlvalue.h"
