+++ src/test/readaheadmanager_test.cpp
@@ -5,7 +5,7 @@
 
 #include "engine/cachingreader.h"
 #include "control/controlobject.h"
-#include "engine/loopingcontrol.h"
+#include "engine/controls/loopingcontrol.h"
 #include "engine/readaheadmanager.h"
 #include "test/mixxxtest.h"
 #include "util/assert.h"
