+++ src/test/signalpathtest.h
@@ -16,7 +16,7 @@
 #include "engine/enginechannel.h"
 #include "engine/enginedeck.h"
 #include "engine/enginemaster.h"
-#include "engine/ratecontrol.h"
+#include "engine/controls/ratecontrol.h"
 #include "engine/sync/enginesync.h"
 #include "mixer/previewdeck.h"
 #include "mixer/sampler.h"
